Resolving compile issues

+++ cloud/src/main/java/com/glolabs/verkundenbot/persistence/repository/IPlugRepository.java
@@ -1,12 +1,12 @@
 package com.glolabs.verkundenbot.persistence.repository;
 
 import java.util.Optional;
-import org.springframework.data.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
 import com.glolabs.verkundenbot.persistence.model.Plug;
 
-public interface IPlugRepository extends JpaRepository<Plug, Long> {
+public interface IPlugRepository extends JpaRepository<Plug, Integer> {
 
-    Optional<Plug> findById(Long id);
+    Optional<Plug> findById(Integer id);
 
     Plug save(Plug plug);
